rag-project/requirements.txt
llama-index
llama-index-core
llama-index-llms-cohere
llama-index-embeddings-cohere
llama-index-vector-stores-pinecone
llama-index-utils-workflow
pinecone-client
python-dotenv
gradio
pathlib
pandas